feat: create todos layout
