+++ app/todo/[year]/[month]/[date]/page.tsx
@@ -1,20 +1,6 @@
-import { cookies } from "next/headers";
-import { redirect } from "next/navigation";
 import { DateParams } from "@/types/params";
-import getPathByDate from "@/utils/getPathByDate";
-import Header from "@/components/DefaultHeader";
-import Calendar from "@/components/Calendar";
 import Todos from "@/components/Todo/Daily";
-import style from "./style.module.scss";
 
 export default function DailyTodosPage({ params }: DateParams) {
-  const path = getPathByDate("todo")({ params });
-  if (!cookies().has("access")) redirect(`/api/login?redirect=${path}`);
-  return (
-    <main className={style.main}>
-      <Header app="todo" {...params} />
-      <Calendar app="todo" {...params} />
-      <Todos {...params} />
-    </main>
-  );
+  return <Todos {...params} />;
 }
